Fix incorrectly generating locations in algorithm
A list comprehension generates almost the right list, but in the wrong
order. The way the list was used, the order must correspond to the order
of Bools in the second list, or else you end up with nonsense moves,
sending requests for unowned sites to move.

I changed it to use the more precise method used in the library
function, but probably a neater way to do things would be to simply
filter the list of sites based on owner, then map `siteLocation` over
the resulting list.